+++ inst/extdata/src/00_main.h
@@ -1,0 +1,5 @@
+#include <cpp11.hpp>
+#include <cpp11armadillo.hpp>
+
+using namespace arma;
+using namespace cpp11;