+++ inst/extdata/src/02_your_functions.cpp
@@ -1,0 +1,3 @@
+#include "00_main.h"
+
+// Add more functions here